+++ src/main/java/org/com/dynamiclantern/client/DynamiclanternClient.java
@@ -2,17 +2,18 @@ package org.com.dynamiclantern.client;
 
 import net.minecraft.world.item.Item;
 import net.minecraftforge.client.ConfigScreenHandler;
+import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.ModList;
 import org.com.dynamiclantern.WaistItemRules;
 import top.theillusivec4.curios.api.client.CuriosRendererRegistry;
 import top.theillusivec4.curios.api.client.ICurioRenderer;
 
+import java.util.HashSet;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
 
 public final class DynamiclanternClient {
-    private static final Set<Item> REGISTERED_RENDERERS = ConcurrentHashMap.newKeySet();
+    private static final Set<Item> REGISTERED_RENDERERS = new HashSet<>();
 
     private DynamiclanternClient() {
     }
@@ -22,6 +23,12 @@ public final class DynamiclanternClient {
         ModLoadingContext.get().registerExtensionPoint(
                 ConfigScreenHandler.ConfigScreenFactory.class,
                 () -> new ConfigScreenHandler.ConfigScreenFactory((minecraft, parent) -> new DynamiclanternConfigScreen(parent)));
+    }
+
+    public static void registerOptionalModListeners(IEventBus modEventBus) {
+        if (ModList.get().isLoaded("epicfight")) {
+            EpicFightWaistItemLayer.register(modEventBus);
+        }
     }
 
     public static void registerConfiguredRenderers() {
@@ -50,8 +57,6 @@ public final class DynamiclanternClient {
     }
 
     private static ICurioRenderer createRenderer() {
-        return ModList.get().isLoaded("epicfight")
-                ? new EpicFightCurioWaistItemRenderer()
-                : new CurioWaistItemRenderer();
+        return new CurioWaistItemRenderer();
     }
 }